+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This update accompanies the release of the [Telegram Bot API 6.7](https://core.t
 
 ### Added
  * Added object `\TgBotLib\Objects\Telegram > InputTextMessageContent` to represent the content of a text message to be sent as the result of an inline query.
+ * Added object `\TgBotLib\Objects\Telegram > InputLocationMessageContent` to represent the content of a location message to be sent as the result of an inline query.